+++ github/__init__.py
@@ -6,4 +6,4 @@ __version__ = "2.0a"
 __license__ = "MIT"
 __copyright__ = "Copyright (c) 2022-present VarMonke & sudosnok"
 
-from .http import HTTPClient
+from .internals import *
